requirements.txt
aiofiles==24.1.0
aiosmtplib==4.0.1
aiohttp==3.12.15
black==25.1.0
fastapi[standard]
httpx==0.28.1
pydantic==2.11.5
pymodbus==3.9.2
pyserial==3.5
python-dotenv==1.1.1
PyYAML==6.0.2
websockets==15.0.1